refactored project for better maintainability
