+++ src/extensions/predictions/PredictionExtension.ts
