+++ src/styles/ErrorOverlay.css
